fix: Control on IDs that do not exist and IDs that are passed as strings in category, role and rolewindow

+++ src/routes/modules/RoleWindows/roleWindows.controller.js
@@ -42,16 +42,38 @@ const roleWindowController = {
     // Get a role-window permission by composite IDs.
     getByIds: async (req, res) => {
         const { idRole, idWindow } = req.params;
+        if (!/^[0-9\s]+$/.test(idRole)) return res.status(400).json({ok: false, error: 'el idde rol solo puede ser numeros'});
+        if (!/^[0-9\s]+$/.test(idWindow)) return res.status(400).json({ok: false, error: 'el idde ventana solo puede ser numeros'});
         try {
             const roleWindow = await roleWindowService.getByIds(idRole, idWindow);
             if(!roleWindow) {
                 return res.status(404).json({
-                    ok: false, error: 'roleWindow no encontrado'
+                    ok: false, error: 'rol ventana no encontrado'
                 });
             }
-            return res.status(200).json({ 
-                ok: true, 
-                data: roleWindow 
+            return res.status(200).json(
+                result
+            );
+        } catch(error) {
+            return res.status(500).json({
+                ok: false,
+                error: 'Error interno del servidor'
+            });
+        }
+    },
+    getByIdRole: async (req, res) => {
+        const { idRole } = req.params;
+        if (!/^[0-9\s]+$/.test(idRole)) return res.status(400).json({ok: false, error: 'el id solo puede ser numeros'});
+        try {
+            const roleWindow = await roleWindowService.getByIdRole(idRole);
+            if(!roleWindow) {
+                return res.status(404).json({
+                    ok: false, error: 'rol ventana no encontrado'
+                });
+            }
+            return res.status(200).json({
+                ok: true,
+                data: roleWindow
             });
         } catch(error) {
             return res.status(500).json({
@@ -61,11 +83,12 @@ const roleWindowController = {
         }
     },
 
+
     // Create a new role-window permission.
     create: async (req, res) => {
         const { idRole, idWindow, create, read, update, remove } = req.body;
         try {
-             if ([idRole, idWindow].some(v => v === undefined)) {
+            if ([idRole, idWindow].some(v => v === undefined)) {
                 return res.status(400).json({ ok:false, error: 'idRole y idWindow son requeridos' });
             }
             const newRoleWindow = await roleWindowService.create({
@@ -93,8 +116,12 @@ const roleWindowController = {
     // Update a role-window permission by composite IDs.
     update: async (req, res) => {
         const { idRole, idWindow } = req.params;
+        if (!/^[0-9\s]+$/.test(idRole)) return res.status(400).json({ok: false, error: 'el id de rol solo puede ser numeros'});
+        if (!/^[0-9\s]+$/.test(idWindow)) return res.status(400).json({ok: false, error: 'el id de ventana solo puede ser numeros'});
 
         const asBool = v => v === true || v === 1 || v === '1' || v === 'true';
+        
+            
 
         const flags = {
             create: asBool(req.body.create),
@@ -105,6 +132,12 @@ const roleWindowController = {
 
         try {
             const updated = await roleWindowService.update(idRole, idWindow, flags);
+            if(!updated) {
+                return res.status(404).json({
+                    ok: false, 
+                    error: 'rol ventana no encontrado'
+                });
+            }
             return res.json({ ok: true, data: updated });
         } catch (error) {
             console.error('[RoleWindow] update error', error);
@@ -115,15 +148,19 @@ const roleWindowController = {
     // Delete a role-window permission by composite IDs.
     delete: async (req, res) => {
         const { idRole, idWindow } = req.params;
+
+        if (!/^[0-9\s]+$/.test(idRole)) return res.status(400).json({ok: false, error: 'el id de rol solo puede ser numeros'});
+        if (!/^[0-9\s]+$/.test(idWindow)) return res.status(400).json({ok: false, error: 'el id de ventana solo puede ser numeros'});
+
         try {
             const deletedRoleWindow = await roleWindowService.delete(idRole, idWindow);
             if(!deletedRoleWindow) {
                 return res.status(404).json({
                     ok: false, 
-                    error: 'role no encontrado'
+                    error: 'rol ventana no encontrado'
                 });
             }
-            res,json({
+            res.json({
                 ok: true,
                 data: deletedRoleWindow
             });
